endrer antallDagerIUken og stillingsprosent ogsa i endre stilling
