+++ src/AvtaleSide/steg/GodkjenningSteg/endringAvAvtaleInnhold/endreTilskudd/EndreTilskuddsberegning.less
@@ -28,14 +28,5 @@
                 flex: 0 45%;
             }
         }
-
-        &__topinputrad {
-            display: flex;
-            justify-content: space-between;
-            width: 100%;
-            align-items: center;
-            margin-top: 0.5rem;
-            padding-right: 2.75rem;
-        }
     }
 }
